refactored routes into route file
