+++ server.js
@@ -1,6 +1,10 @@
 const express = require("express");
 const path = require("path");
 const bodyParser = require("body-parser");
+const passportJWT = require("passport-jwt");
+const passport = require("passport");
+const routes = require("./routes/routes.js");
+const auth = require("./auth");
 
 const app = express();
 
@@ -10,7 +14,16 @@ app.use(bodyParser.urlencoded({
 
 app.use(bodyParser.json());
 
-require('./auth')(app);
+app.use(passport.initialize());
+
+var jwtOptions = {
+    jwtFromRequest: passportJWT.ExtractJwt.fromAuthHeaderAsBearerToken(),
+    secretOrKey: 'zse45tgb'
+  }
+
+auth(passport, passportJWT, jwtOptions);
+
+routes(app, passport, jwtOptions);
 
 app.use('/',express.static('build'));
